# docs/swagger.yaml
basePath: /api
definitions:
  github_com_pozgo_web-cli_internal_models.BashScriptCreate:
    properties:
      content:
        type: string
      description:
        type: string
      filename:
        type: string
      group:
        description: Optional, defaults to "default"
        type: string
      name:
        type: string
    required:
    - content
    - name
    type: object
  github_com_pozgo_web-cli_internal_models.BashScriptResponse:
    properties:
      content:
        description: Only included when specifically requested
        type: string
      created_at:
        type: string
      description:
        type: string
      filename:
        type: string
      group:
        description: Group/category for organization
        type: string
      id:
        type: integer
      name:
        type: string
      source:
        description: '"sqlite" or "vault"'
        type: string
      updated_at:
        type: string
    type: object
...
  internal_server.ErrorResponse:
    description: Error response returned by the API
    properties:
      error:
        example: Invalid request body
        type: string
    type: object
...
host: localhost:7777
info:
  contact:
    email: linux@ozgo.info
    name: Web CLI Support
    url: https://github.com/pozgo/web-cli
  description: Basic authentication using username and password
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Web CLI API
  version: 1.1.0
paths:
  /bash-scripts:
    get:
      consumes:
      - application/json
      description: Get a list of all bash scripts (without content by default)
      parameters:
      - description: Filter by group name
        in: query
        name: group
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.BashScriptResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: List all bash scripts
      tags:
      - Bash Scripts
    post:
      consumes:
      - application/json
      description: Create a new bash script (stored encrypted)
      parameters:
      - description: Bash script to create
        in: body
        name: script
        required: true
        schema:
          $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.BashScriptCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.BashScriptResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Create a bash script
      tags:
      - Bash Scripts
  /bash-scripts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a bash script by its ID
      parameters:
      - description: Bash Script ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Delete a bash script
      tags:
      - Bash Scripts
    get:
      consumes:
      - application/json
      description: Get a single bash script by its ID
      parameters:
      - description: Bash Script ID
        in: path
        name: id
        required: true
        type: integer
      - default: true
        description: Include script content in response
        in: query
        name: include_content
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.BashScriptResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Get a bash script by ID
      tags:
      - Bash Scripts
    put:
      consumes:
      - application/json
      description: Update an existing bash script by its ID
      parameters:
      - description: Bash Script ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bash script update data
        in: body
        name: script
        required: true
        schema:
          $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.BashScriptUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.BashScriptResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Update a bash script
      tags:
      - Bash Scripts
  /bash-scripts/{id}/presets:
    get:
      consumes:
      - application/json
      description: Get all presets for a specific bash script
      parameters:
      - description: Bash Script ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.ScriptPresetResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Get presets for a script
      tags:
      - Script Presets
  /bash-scripts/execute:
    post:
      consumes:
      - application/json
      description: Execute a stored bash script locally or remotely
      parameters:
      - description: Script execution request
        in: body
        name: execution
        required: true
        schema:
          $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.ScriptExecution'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.ScriptResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Execute a bash script
      tags:
      - Bash Scripts
  /bash-scripts/execute/stream:
    post:
      consumes:
      - application/json
      description: Execute a stored bash script locally or remotely with real-time
        output streaming via SSE
      parameters:
      - description: Script execution request
        in: body
        name: execution
        required: true
        schema:
          $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.ScriptExecution'
      produces:
      - text/event-stream
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server.StreamMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Execute a bash script with streaming output
      tags:
      - Bash Scripts
  /bash-scripts/groups:
    get:
      consumes:
      - application/json
      description: Get a list of all distinct group names for bash scripts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: List all bash script groups
      tags:
      - Bash Scripts
  /commands/execute:
    post:
      consumes:
      - application/json
      description: Execute a shell command locally or remotely via SSH
      parameters:
      - description: Command execution request
        in: body
        name: command
        required: true
        schema:
          $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.CommandExecution'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.CommandResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Execute a command
      tags:
      - Commands
  /env-variables:
    get:
      consumes:
      - application/json
      description: Get a list of all environment variables (values masked by default)
      parameters:
      - description: Show actual values instead of masked values
        in: query
        name: show_values
        type: boolean
      - description: Filter by group name
        in: query
        name: group
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.EnvVariableResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: List all environment variables
      tags:
      - Environment Variables
    post:
      consumes:
      - application/json
      description: Create a new environment variable (stored encrypted)
      parameters:
      - description: Environment variable to create
        in: body
        name: envvar
        required: true
        schema:
          $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.EnvVariableCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.EnvVariableResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Create an environment variable
      tags:
      - Environment Variables
  /env-variables/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an environment variable by its ID
      parameters:
      - description: Environment Variable ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Delete an environment variable
      tags:
      - Environment Variables
    get:
      consumes:
      - application/json
      description: Get a single environment variable by its ID
      parameters:
      - description: Environment Variable ID
        in: path
        name: id
        required: true
        type: integer
      - description: Show actual value instead of masked value
        in: query
        name: show_value
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.EnvVariableResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Get an environment variable by ID
      tags:
      - Environment Variables
    put:
      consumes:
      - application/json
      description: Update an existing environment variable by its ID
      parameters:
      - description: Environment Variable ID
        in: path
        name: id
        required: true
        type: integer
      - description: Environment variable update data
        in: body
        name: envvar
        required: true
        schema:
          $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.EnvVariableUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.EnvVariableResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Update an environment variable
      tags:
      - Environment Variables
  /env-variables/groups:
    get:
      consumes:
      - application/json
      description: Get a list of all distinct group names for environment variables
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: List all environment variable groups
      tags:
      - Environment Variables
  /history:
    get:
      consumes:
      - application/json
      description: Get command execution history with optional filtering
      parameters:
      - description: Filter by server name
        in: query
        name: server
        type: string
      - default: 100
        description: Maximum number of records to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.CommandHistory'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: List command history
      tags:
      - Command History
  /history/{id}:
    get:
      consumes:
      - application/json
      description: Get a single command history entry by its ID
      parameters:
      - description: Command History ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.CommandHistory'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Get a command history entry by ID
      tags:
      - Command History
  /keys:
    get:
      consumes:
      - application/json
      description: Get a list of all SSH keys stored in the system
      parameters:
      - description: Filter by group name
        in: query
        name: group
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.SSHKey'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: List all SSH keys
      tags:
      - SSH Keys
    post:
      consumes:
      - application/json
      description: Store a new SSH private key in the system
      parameters:
      - description: SSH Key to create
        in: body
        name: key
        required: true
        schema:
          $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.SSHKeyCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.SSHKey'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Create a new SSH key
      tags:
      - SSH Keys
  /keys/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an SSH key by its ID
      parameters:
      - description: SSH Key ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Delete an SSH key
      tags:
      - SSH Keys
    get:
      consumes:
      - application/json
      description: Get a single SSH key by its ID
      parameters:
      - description: SSH Key ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.SSHKey'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Get an SSH key by ID
      tags:
      - SSH Keys
    put:
      consumes:
      - application/json
      description: Update an existing SSH key by its ID
      parameters:
      - description: SSH Key ID
        in: path
        name: id
        required: true
        type: integer
      - description: SSH Key update data
        in: body
        name: key
        required: true
        schema:
          $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.SSHKeyUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.SSHKey'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Update an SSH key
      tags:
      - SSH Keys
  /keys/groups:
    get:
      consumes:
      - application/json
      description: Get a list of all distinct group names for SSH keys
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: List all SSH key groups
      tags:
      - SSH Keys
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: 'Bearer token authentication (format: "Bearer {token}")'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...